feat(#87): add support gacha banner metadata
